Moved BugSplat class into its own file. Added DS_Store to .gitignore.
