+++ .gitignore
@@ -2,3 +2,4 @@
 **/__pycache__
 build
 dist
+.DS_Store
